PCIDMA: Add PCI DMA performance test driver
When PCIE controller working in EP mode, it can be connected to
a host such as x86 computer or PowerPC board as a PCI device.
Here is the diagram for hardware setup and software components.

          RC (X86-64/PowerPC)            EP (PowperPC card)
	-----------------------        -----------------------
	|                     |        |                     |
	|   PCI DMA driver    |        | PCI EP application  |
	|                     |  PCIe  |                     |
	|   PCIe data path    |<======>|   PCIe data path    |
	-----------------------        -----------------------

This patch is PCI DMA driver running in the host side and is used
to measure PCI memory write performance. It supports multiple PCI
functions and can create a thread on the specified function to
measure performance using host DMA or memcpy. It also can control
EP device to start EP DMA and calculate the performance. The driver
also provides some sysfs interface to change test settings such as
packet length loop times and write direction(RC to EP or EP to RC).

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>
Signed-off-by: Liu Gang <Gang.Liu@freescale.com>
Signed-off-by: Mingkai Hu <Mingkai.Hu@freescale.com>