PCIDAM: Use streaming mapping instead of coherent mapping
1. For DMA data transferring, the patch uses streaming DMA mapping
instead of coherent DMA mapping. There are two reasons. The first
is that, on systems that support mapping registers, each DMA mapping
uses one or more of them on the bus. Coherent mappings, which have a
long lifetime, can monopolize these registers for a long time, even
when they are not being used. The other reason is that, on some
hardware, streaming mappings can be optimized in ways that are not
available to coherent mappings.
2. The patch assigns different values to DMA memory region for it
is convenient to verify reading/writing results.
3. If info is NULL, pcidma_test_info_free() directly returns to
avoid crashing

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>