fix dequeueing problem for dsa sign-verify test
The sign-verify test is a two part test: a sign and a verify test. The
verify test is started in the completion callback of the sign test. If
the verify job cannot be enqueued, for example because the input queue
is full, the enqueueing function returns -1 which is never used to
decrease the enqueued job counter causing a mismatch between the number
of the enqueued and dequeued jobs that prevents the test from stopping.
To fix this problem the job enqueued counter is decremented whenever the
verify job cannot be enqueued.

Signed-off-by: Radu Alexe <raduandrei.alexe@freescale.com>